Updated app.py
Updated app.py. We submitted the project deliverables on November 19 itself, but we unknowingly added the SendGrid mail integration API Key in the GitHub repository. So, SendGrid told us to delete the API Key. After deleting the API Key it shows the error in our project while executing. So, again we have created a new API Key and deployed the project and now we are adding the updated file.